rename kebab-case files
